requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
redis==4.6.0
channels==4.1.0
channels-redis==4.1.0
daphne==4.1.0
asgiref
black==23.1.0
flake8==6.0.0
drf_spectacular
bson
Pillow
